Added new expression
A new expression had to be unfortunately added to correctly handle the enumerations. This expression, called EnumVariantExpression, is used to assign values to an enumerate.

Also, I have modified the fieldValueAssignmentsExpression parser to adapt it to put the type at the end, followed by a ":", as with the rest of the typed expressions.