Changed the AST to fix the Undyn object expression
Now the Undyn are object access expressions instead of regular expressions. The
following changes have been applied to the code:

1) The data type Expression has been moved to CoreAST, changing the name to
Expression'.
2) The new data type Expression' is parametric on the object data type.
3) Now both AST and SemanAST implement their own object data type which in turn
depends on Expression 🙃.
4) Both data types only receive the type of Identifier Expressions and the type
of annotations. The type of the expressions is fixed to Expression (Expression
is a type alias of (Expression' RHSObject)).
5) Changed references to the objects in both the typechecker and the parser to
make the typechecker work.

The src/Utils/CoreAST.hs file has been maintained despite being empty so that
changes can be traced.