Added task printer and initial state
The initial state has been completed with a set of defined functions and types that would be part of the standard Termina library and runtime. Specifically, the functions are part of the runtime itself and are not coded in Termina. In the future we would have to see how to incorporate the inclusion of the standard library in a more elegant way.

The task printer has also been incorporated. More testing is still to be done.