Removed parenthesis expressions from the AST
We have eliminated the expressions in parentheses as part of the AST. Now, parentheses are printed by the pretty printer only when they are needed depending on the precedence of the different expressions and operators. The precedence of operators is the same as in C.

This change allows eliminating redundant parentheses from the generated code, thus increasing its quality.

All tests pass.