Change in the definition of arrays
We have made the following changes:

- The definition of arrays has been changed so that the dimension is an integer constant with no type. This makes it necessary to recode existing code to remove the type from the dimension.
- A new unsigned integer type named `usize` (USize) has been added (à la Rust). This type is equivalent, and is transpiled directly, to the `size_t` type in C. This type is an unsigned integer of size equal to the size of the target machine addresses.
- The new type usize is the type now used to index all arrays and to create the slices. It is necessary to recode the existing code to adapt it to the new situation.

Currently, the bit size of the `usize` type is fixed at 32 bits, although the type itself is transpiled to `size_t`. In the future, it will be necessary to tell the transpiler the size of the `size_t` type for the target machine so that it can check the definition of the constants of that type.