Updated match statement pretty printing
The code of the match statement cases with associated parameters was not generated correctly. In non-anonymous match statements, i.e. where the expression was an object, the case accessed the object fields directly instead of making a copy. This could cause a problem if the value of the original object was overwritten. Now, in the case of non-anonymous match statements, a copy is made when entering the corresponding case.

All tests pass.