Added DereferenceMemberAccess element
The DereferenceMemberAccess AST element has been added, corresponding to the `->` operator used to call the methods or viewers of a class from a reference. This operator is only used to call the methods of the class itself from its `self` reference, since in the case of procedure calls, they are made through a port.

The type checker code has been minimally restructured to avoid duplicity in the code.

New tests have been added to test the new AST element.