Modified file printing functions
The functions now receive the path of the current module being printed and in addition, in the case of header files, the list of paths of the modules on which it depends.

The generated code incorporates the necessary include directives for successful compilation.

The tests have been modified to take into account the new transpiler output.

All tests pass.