Minor modifications to the message queues API
The `send` function now receives a mutable reference to a `Result` as a second parameter. This way, the user code can know if any error has occurred in sending the data (i.e., if the message queue was full at the time of sending).

The change in the `Result` and `TaskRet` types has been reverted. It returns to the original `Error` variant, i.e., without parameter. It was felt that this adds simplicity and that, in the vast majority of cases, there is only a single cause of error.