Generation of the initialization function of global objects
First version of the global object initialization function generator. This function is independent of the underlying operating system. The function is responsible for initializing the state of the various global objects. This function must be called by the runtime ONLY ONCE before the different tasks are launched and the interrupt handlers can be executed.

In order to properly implement this function, it has been necessary to modify the AST to add type annotations in the assignment of `Location` and `Port` type fields.

A small code restructuring has also been performed. Future reorganization and refactoring of the printing functions, which are now split between the `PPrinter` directory and the `Modules` directory, will be necessary.