Minor changes
In this new version, the code generation of RTEMS tasks has been modified. Now, instead of having one function per task, only one function per class is defined. The ability of RTEMS to pass an argument to the task is used to pass as a parameter the address of the corresponding structure.