Fixed a bug in the access to dyn structure fields
Now, when accessing the field of a structured type variable, it is first checked if the variable is of dynamic type. If so, an undyn is applied to the variable before accessing the field.

One of the integration tests has been modified to test the correctness of the bug.

All tests pass.