Added support for non-dynamic Options
Until now, the Option type was not really polymorphic: it could only be used with dynamic subtypes (e.g., Option<dyn u32>. Now, the Option type is fully polymorphic: you can define an Option of any simple type. To make this new implementation work, the transpiler has been modified to generate a structure for each use of the polymorphic type that can be made. For example, *for each of these types*:

- `Option<u32>`
- `Option<[u32; 10]>`
- `Option <[[u32; 10]; 10]>`

a structure containing two fields is defined: one to store the value corresponding to the `Some` variant and another with the enum indicating whether the value is valid (`Some`) or not (`None`). These structures are defined on-demand only. Options of non-dynamic types can be used as any other simple type, that is, to declare variables, input parameters and fields of structures and classes. It is necessary to review the type checker code to check that all the present rules are still in effect and that the new ones are implemented.

This new feature implies a modification in the use of message queues. Message queues can now receive dynamic objects or simple objects. In the case of receiving simple objects, the object is copied to the message queue and, therefore, the ownership of the object is not lost.

The change has required the modification of several points of the code, i.e.,

- Once the modules are typed, and after doing the DF check, the code of all the modules is traversed generating a map containing the Options defined for each type. This map, in turn, is partitioned into two: one for the basic types (integer, char and bool), and another for the user-defined types. In order to avoid duplicates and not to generate the same structure multiple times, the elements of the map are sets of types. The code that searches for Option types is located in `src/Semantic/Option.hs`.
- The printer of the type definition code has been modified: if the code incorporates the use of an Option containing a user-defined type (struct or enum), the transpiler will generate the corresponding structures after those of the new type. To know if it has to generate any Option structure for a given type, the `ppTypeDefDeclaration()` function receives the map with the Option types.
- The type checking of the `send` procedure of message queues has been modified to force the first parameter to always be an object. This is necessary because, being a polymorphic procedure, the resulting function in C actually receives the object by reference, that is, although in termina the object is not passed by reference, in C it is. Therefore, it is not possible to pass as parameter any expression, since the operator `&` in C does not allow to obtain the addresses of anonymous objects, e.g.: `send(&get_integer())` is not valid in C (nor in termina).
- Options can now be of any simple type except an Option type, i.e., nested Options are not allowed, e.g. `Option<Option<u32>>` is not a valid type.

The tests have been modified to adapt them to the new definitions. Further testing is needed to fully validate the new implementation and to check that the other components of the transpiler (type checker and DF) behave as expected.

All tests pass.