Inclusion of new language elements
We have added new elements to the language: emitters, channels, actions and
interfaces. The emitters are global objects managed by the runtime that
generate events to be handled by tasks and handlers. Channels are global
objects, such as message queues, that allow the exchange of messages between
tasks and handlers.

Tasks and handlers can associate actions to the arrival of a certain event.
These actions are defined as internal functions of tasks and events.  Four
different types of ports may been defined: access, input, output and event sink
ports. Access ports are the old resource ports. These ports now have an
associated interface instead of directly referencing a resource class. These
interfaces are explicitly defined as a set of procedures. Resources can provide
one or more of these interfaces.  Right now, we do not check (big TODO) that
interfaces are correctly implemented by resources nor that a resource connected
to a port implements the interface associated to the port. In addition to the
user-defined interfaces, there is a polymorphic interface defined within the
language itself called Allocator<T>, which is the interface provided by the
memory pools. This interface has a type parameter T corresponding to the type
of the elements served by the pool.

In addition to this main change, the following changes have been incorporated:

- The free function is now a procedure of the pool. This change forces the user
  to explicitly know and state from which pool the dynamic memory he is
  receiving has originated and to establish the necessary connections between the
  memory receiver elements and the pools.

- Elimination of message reception functions. Now it is the runtime that
  directly extracts the messages from the queues and calls the task actions
  when appropriate. In this way, we can implement the run-to-completion semantics
  of the tasks.

- The message sending function only receives as parameter the data to be sent.
  Previously it also received a second parameter corresponding to the passing
  by reference of a variable that stored the result of the sending. Now,
  unsuccessful sending is treated as a system event to which a handler can be
  associated.

- Modification of the types of the field mapping expressions used to connect
  resources. The types of these expressions now contain the class of the
  resource being connected and the list of procedures belonging to the interface
  provided by the resource. This list is then used by the pretty printer to
  generate the initialization code.