Multiple changes
The following changes are included with this commit:

- Completed the first alpha version of the main application code generator for
  RTEMS on NOEL. The generated code has not yet been tested nor compiled. One
of the substantial changes is that the tasks and handlers contain a copy of the
message queue identifiers corresponding to the incoming and outgoing ports. In
this way, we fully control the accessed memory addresses (this may be relevant
for PREM deployments).

- Added a new operator, "is", which allows to check if the value of the
  enumerated type object corresponds to a given variant. The resulting
expression is of boolean type. The operator does not destroy the object: it
only returns "true" if the object variant corresponds to the specified one,
e.g.:

```
var res : Result = Result::Ok;
...
if (res is Result::Error) { ... }
```

The operator does not allow access to the possible data fields associated with
the variant. It works for both enumerations and option objects.

- The naming of many internal structures has been homogenized.

- The pretty printer of the functions has been modified so that the name of the
  function can be replaced by some identifier of the substitution map. This
change has been included mainly to be able to rename built-in functions (e.g.
clock_get_uptime) to their corresponding runtime functions (e.g.,
__termina__clock_get_uptime).

- The tests have been modified to adapt them to the new naming. A new test has
  been defined for the "is" operator. All tests pass.