First version of the new code generator
This commit includes the first version of the new code generator. This
generator is strongly inspired by the Haskell language-c package. Now, the
generator does not directly write the statements as they are to be printed, but
first generates a C AST which is then printed. The code generator is still to
be finished and the existing tests are still to be adapted. At the moment, only
the IT of the arithmetic expressions has been updated.