Added declaration statements
Support for the generation of declaration statements using the new
infrastructure has been added. We have added new types of annotations to
facilitate code generation in C. In particular, annotations have been added
that allow you to modify the behavior of the code generator to add or not an
additional line before a declaration or statement.