Completed the generation of the rest of the elements
We have finished implementing the generator for the rest of the elements
(classes, functions, modules and global variables). More testing is still
needed. Some of the tests have been adapted, but the rest are yet to be
included.

In addition, the generation of the application code still needs to be
reimplemented.