Initial commit of the RTEMS NOEL glue generator
The generator is not yet completed.

A new parameter has been added to the statement annotations. This new parameter
allows you to specify whether the pretty printer should indent a statement when
printing it or not. This parameter allows indenting the statements that appear
after a statement of type Case inside a Switch.