First version of the generator completed
The new code generator, which uses a translation to a C AST before printing the
code, now implements the same functionality as the original pretty printer.
Until the test code is properly tested, we will not remove the original version
from the repository.

The tests have been corrected by adding a CR at the end of each line. This CR
is added by the code generator to make sure that all files have a trailing new
line.