Initial version of const generics
The commited code is not yet fully functional. Right now, changes have been
completed on the AST so that functions, procedures and viewers can receive
const generics parameters. The code generator has also been adapted to generate
the corresponding C function parameters. A lot of testing still needs to be
done and the implementation of the type checker still needs to be completed.

The AST has also been modified to annotate integers with a parameter
corresponding to their representation (decimal or hexadecimal). This parameter
is then used to generate the C constants and print the corresponding literals.

All unit and integration tests have been adapted. All tests pass (although none
of them tests the new functionality yet).