Improved type checker
The type checker is now able to infer from the context the type of integer
literals. The error message mechanism needs to be MUCH improved.

The type checker, in any case, if it detects any ambiguity in the type of an
integer literal, generates an error. In such cases, the user must provide the
explicit type of the literal.

Much remains to be tested and implemented.