Modified use of constant expressions
Now the limits of for loops must be constant expressions, that is, integer
literals or global constants. In addition, all input arguments of functions
become immutable. This has led to the need to modify a number of poorly defined
integration tests.

Global constants are transpiled as defines. The code generator uses macros
to ensure the size of constants (with fixed width integer types). It remains to
be seen whether this makes sense or not and whether it would be necessary to
extrapolate its use to all integer literals in the program.