Minor modification
We have applied a small modification that allows us to detect when there is or
is not an else block in a conditional statement. This allows to generate an
error when, after an if-else-if statement, there is no last else block. The
need for an Else block is derived from MISRA-C guidelines. This modification
resolves issue #50.

We have added a new test to validate that the condition is detected correctly.