First tests with const generics
We have implemented the first tests of the function definition using const
generics. We have made small modifications in the type checker and corrected
small bugs (e.g. the type of the function parameters was not checked) to
support the definition of functions with const generics parameters.

We have implemented some small tests to validate the implementation.