Initial commit of the new pretty printer error
We have modified the compiler to print the errors using the Errata library.
This library allows to generate text blocks with the errors in a simple way. In
order to use it, we had to incorporate the content of the files into the
structures used by the transpiler. In this way, we can print the line where the
error is found.

We have added only a few sample errors. It will be necessary to add new errors.