Parameter check
We have modified the type checker code so that it verifies that the parameter
types of functions and procedures are valid. We have changed the specification
of valid parameter types to NOT ALLOW functions to define parameters or return
array values. Thus, we can only pass references to arrays, either mutable or
immutable, as parameters to functions. This has resulted in the removal of
several tests that defined functions that return and receive arrays.

The fact that functions can only receive arrays allows us to overload the
indexing operator. This way, the operator can be used directly on references
without having to dereference the reference to the arrays.

We have added new tests and new cases to the error printer.

We have eliminated spurious imports in some files.