Fixed bug in semantic parser
We have fixed a bug in the semantic analyzer. In fact, we have applied several
modifications to the dependency detector and to the topological order generator
from the dependency graph in order to correctly trace the errors. The applied
change is likely to be too costly, but we have chosen to modify as little as
possible the algorithm that obtains the order, since it is used elsewhere in
the transpiler code. Now, when it is time to obtain the dependencies between
the members of the classes, a map is generated with the points where these
dependencies are found. In this way, thanks to this map, it is possible to
correctly trace the source of the error and detail the exact point where the
reference has not been found.