Added more inference capabilities
The type checker now allows you to use a constant on one side of a binary
expression as long as you can infer its type from the other operand.

In addition, a minor modification has been made to the type checker that allows
the type of a struct initializer to be omitted, since it can always be inferred
from the environment.