Minor modification of AtomicArrayAccess interface
Now the interface also forces to define the size of the atomic array. This is
because the code of the entity using the resource must know the size of the
array. It cannot change based on the environment, such as the size of the
message queues. It must be hardcoded.