Major name change
We have changed the name of the subtype “dyn” to “box”. This name better reflects the semantics of the type and allows us to relate it to equivalent types in other programming languages such as Rust.

We have updated all names and comments.

All tests pass.

All documentation remains to be modified.