Initial Commit of the new C language AST
This commit contains the first version of the new C language AST. This AST is
very similar (but not the same) to the CompCertC AST. This change has been
introduced with the idea that, in the future, we can implement the transpiler
using Coq (or similar) and be able to formally verify that it is correctly
implemented according to the operational semantics of Termina.

We have had to make minor changes to the Termina AST to remove the Slice type,
which conflicted with the new C language AST. The transpiler is not yet fully
reimplemented and the C expression typing is most likely not well done. There I
leave it.