Initial commit of the new C templating mechanism
The new C templating mechanism is intended to helping the generation of the glue code. It basically embeds a subset of the C language in Haskell so that the code specification is done much more simpler and easier.

Also, a new expression (size_of(expr)) has been added.

The glue code generator for the platform has not been finished and will be commited next.