Minor changes
We have modified the definition of the Procedures to homogenize the code blocks. Now they are all of the BlockRet class. This will allow reusing the central core of the language by moving from blocks defined as a sequence of statements to blocks defined as a list of basic blocks.