Major code reorganization
We have performed a major code reorganization. There is no longer an AST
directory with the ASTs, but each AST is in its own directory (Core, Parser,
Seman, etc.).

Also, now the main AST does not define how the code blocks have to look like.
This will allow in the near future to be able to define ASTs that implement the
code as basic block lists.