Initial commit of the control flow AST
This AST implements a syntax in which the code blocks are represented as a
sequence of basic blocks. The objective is that the AST generated by the type
checker is translated into this new AST, so that it serves as input for the
code generator on the one hand and the static analyzer(s) on the other.