Updated block definition
We have eliminated the requirement that the return statement be the last
statement. Now, it can happen that, with the new continuations, the code of an
action delegates the execution to another internal action.

All tests pass. It is now necessary to manually check the blocks to make sure
that they are well formed. This check must be done after running the type
checker.