Private objects are back
After a friendly discussion with team members, we have decided to re-implement private objects. Now, task fields and handlers are private objects. The purpose of this is to force the programmer to use as task fields and handlers only those simple variables that only represent the state of the active object in question. The object data will be stored in resources external to the task. When these resources are not shared by any other task, the data is not protected.

In addition to this change, we have defined new errors and increased the number of errors that are pretty printed.