Initial implementation of the box source checker
We have now completed the initial implementation of the box source checker. It
is in charge of checking that the boxes are always released in the pool from
which they are extracted. To do so, it uses the box source map of each class
generated during the application architecture analysis phase. These maps relate
the box outputs to the inputs. The checker is in charge of verifying all
possible paths leading to each pool.