Refactoring of the usage checker code
We have homogenized the implementation of different functions and made a first
implementation of the pretty printer of the errors. It is necessary to redefine
some errors in order to display more complete information to the user.

Several bugs have been corrected. More tests are needed to verify that the
checker works correctly. Now, the checker allows to reallocate memory in a
variable of type option-box that has been previously used.