Fixed a problem with the type definition
When using parametric types (such as atomic arrays or message queues), and the size parameter was specified as a constant, the program did not work correctly. This was because the parser interpreted the second parameter as a type, not as a size parameter.

Now, the parsing of identifiers is done separately and they are treated as types or sizes (constants) depending on the position of the parameter.