Fixed issue #58
The original code contained an error. In the original implementation, the type checking function sorted both the case identifiers and the enumerated variant identifiers alphabetically. This sorting was performed regardless of whether the match statement included the default case or not. If the list of cases was exhaustive (i.e., there was no default case), the program worked correctly, as all cases were checked one by one and compared with the variants, both sorted alphabetically.

The problem occurred when the statement defined a default case. In that case, the check continued to be performed using the alphabetically sorted lists. The problem arose when the identifier of a case was checked and its position did not correspond to that of the original list of variants. In that case, when comparing the case identifier with the variant identifier, a mismatch occurred erroneously.

We have updated the code to fix the bug. The errors detected in issue #58 have been fixed. All tests pass.