Major code change
We have made a substantive change in the way actions are implemented. Now, regular actions return an object of type Status<i32>. This is one of the two new monadic types we have defined. The Status<T> type has two variants: Success and Failure(T). The first one is used to indicate that the computation was successful. The second is used to indicate that a failure has occurred. The data associated with the Failure variant indicates the type of data that has occurred.

The second monadic type is Result<L; R>, which has two variants: Ok(L) and Error(R). The first variant is used to indicate that the computation has completed successfully, returning a value of type L. The second is used to indicate that the computation has been erroneous, and as a result, an error code of type R has been generated.

Both monadic types can be used at the discretion of the programmer.

To support the new actions, we have had to apply changes to the OSAL as well.