Added possibility to use references in match statements
Now it is possible to use references in match statements. In case of using them, the arguments are taken as references to the changes of the original structure. In this way, we can modify the contents of the object directly, without making a copy of the parameters.

Option types are not yet supported.