Added support for Option references in match statements
Now, you can pass a reference to an Option object in a match statement. This makes the memory stored in the Option directly accessible instead of being copied. In addition, several bugs related to variable usage checking and procedure code generation have been fixed.

In the former case, a problem concerning references to Option-box objects has been solved. Previously, making the reference directly moved the object instead of borrowing it.

In the latter case, the definition of the variable used to store the previous state of the interrupts when locking and unlocking could cause a name collision with the procedure parameters.