feat: added emulated keyboard interrupt in POSIX system support
In this new version of the software, we provide support for an emulated interrupt in POSIX systems. The OSAL is capable of generating an interrupt (active by level) every time there is new data on the keyboard. In addition, we have added a new system call, read(), which allows data to be read from standard input in a non-blocking manner. This new implementation will allow the behavior of an interrupt-based system to be emulated and more realistic POSIX demonstrators to be implemented.

The read() system call is not available for the RTEMS platform.